Convert more x-pack modules to new test framework (#125874)
(cherry picked from commit 421392af99f078798df762c4e9b78b3ea86b7457)

# Conflicts:
#	x-pack/plugin/slm/qa/multi-node/build.gradle
#	x-pack/plugin/slm/qa/with-security/build.gradle
#	x-pack/plugin/stack/qa/rest/build.gradle
