+++ x-pack/plugin/slm/build.gradle
@@ -2,15 +2,18 @@ import org.elasticsearch.gradle.internal.info.BuildParams
 
 apply plugin: 'elasticsearch.internal-es-plugin'
 apply plugin: 'elasticsearch.internal-cluster-test'
+apply plugin: 'elasticsearch.internal-yaml-rest-test'
+apply plugin: 'elasticsearch.internal-java-rest-test'
 
 esplugin {
   name = 'x-pack-slm'
   description = 'Elasticsearch Expanded Pack Plugin - Snapshot Lifecycle Management'
-  classname ='org.elasticsearch.xpack.slm.SnapshotLifecycle'
+  classname = 'org.elasticsearch.xpack.slm.SnapshotLifecycle'
   extendedPlugins = ['x-pack-core']
-  hasNativeController =false
-  requiresKeystore =true
+  hasNativeController = false
+  requiresKeystore = true
 }
+
 base {
   archivesName = 'x-pack-slm'
 }
@@ -21,6 +24,20 @@ dependencies {
   testImplementation project(xpackModule('ccr'))
   testImplementation project(xpackModule('ilm'))
   testImplementation project(':modules:data-streams')
+  javaRestTestImplementation(testArtifact(project(xpackModule('core'))))
+  javaRestTestImplementation project(xpackModule('slm'))
+  yamlRestTestImplementation(testArtifact(project(xpackModule('core'))))
+  clusterModules project(xpackModule("ilm"))
+  clusterModules project(xpackModule("searchable-snapshots"))
+  clusterModules project(":modules:data-streams")
 }
 
-addQaCheckDependencies(project)
+restResources {
+  restApi {
+    include '_common', 'cluster', 'indices', 'index', 'snapshot', 'slm', 'health_report'
+  }
+}
+
+tasks.named("javaRestTest") {
+  usesDefaultDistribution("uses _xpack/usage api")
+}
