--- x-pack/plugin/slm/qa/multi-node/build.gradle
@@ -1,37 +1,0 @@
-import org.elasticsearch.gradle.internal.info.BuildParams
-import static org.elasticsearch.gradle.PropertyNormalization.IGNORE_VALUE
-
-apply plugin: 'elasticsearch.legacy-java-rest-test'
-
-dependencies {
-  javaRestTestImplementation(testArtifact(project(xpackModule('core'))))
-  javaRestTestImplementation project(xpackModule('slm'))
-}
-
-File repoDir = file("$buildDir/testclusters/repo")
-
-tasks.named("javaRestTest").configure {
-  /* To support taking index snapshots, we have to set path.repo setting */
-  nonInputProperties.systemProperty 'tests.path.repo', repoDir
-}
-
-testClusters.configureEach {
-  testDistribution = 'DEFAULT'
-  numberOfNodes = 4
-
-  setting 'path.repo', repoDir.absolutePath, IGNORE_VALUE
-  setting 'xpack.searchable.snapshot.shared_cache.size', '16MB'
-  setting 'xpack.searchable.snapshot.shared_cache.region_size', '256KB'
-  setting 'xpack.security.enabled', 'false'
-  setting 'xpack.watcher.enabled', 'false'
-  setting 'xpack.ml.enabled', 'false'
-  setting 'xpack.license.self_generated.type', 'trial'
-  setting 'indices.lifecycle.poll_interval', '1000ms'
-  setting 'logger.org.elasticsearch.xpack.core.slm', 'TRACE'
-  setting 'logger.org.elasticsearch.xpack.slm', 'TRACE'
-}
-
-if (buildParams.inFipsJvm){
-  // Test clusters run with security disabled
-  tasks.named("javaRestTest").configure{enabled = false }
-}
