+++ x-pack/plugin/slm/src/javaRestTest/java/org/elasticsearch/xpack/security/PermissionsIT.java
