+++ x-pack/plugin/slm/src/yamlRestTest/resources/rest-api-spec/test/slm/11_basic_slm.yml
