+++ x-pack/plugin/slm/src/yamlRestTest/resources/rest-api-spec/test/slm/20_health.yml
