--- x-pack/plugin/stack/qa/rest/build.gradle
@@ -1,23 +1,0 @@
-apply plugin: 'elasticsearch.legacy-yaml-rest-test'
-apply plugin: 'elasticsearch.legacy-yaml-rest-compat-test'
-apply plugin: 'elasticsearch.authenticated-testclusters'
-
-dependencies {
-  yamlRestTestImplementation(testArtifact(project(xpackModule('core'))))
-}
-
-restResources {
-  restApi {
-    include '_common', 'cluster', 'indices', 'index', 'snapshot', 'ilm', 'slm', 'stack', 'indices'
-  }
-}
-
-testClusters.configureEach {
-  testDistribution = 'DEFAULT'
-  setting 'xpack.ml.enabled', 'false'
-  setting 'xpack.license.self_generated.type', 'trial'
-}
-
-tasks.named("yamlRestTestV7CompatTransform").configure({ task ->
-  task.skipTest("stack/10_basic/Test kibana reporting index auto creation", "warning does not exist for compatibility")
-})
