+++ x-pack/plugin/vector-tile/build.gradle
@@ -13,7 +13,7 @@
  */
 
 apply plugin: 'elasticsearch.internal-es-plugin'
-apply plugin: 'elasticsearch.legacy-java-rest-test'
+apply plugin: 'elasticsearch.internal-java-rest-test'
 
 esplugin {
   name = 'vector-tile'
@@ -34,12 +34,11 @@ dependencies {
   //  runtimeOnly("org.apache.logging.log4j:log4j-slf4j-impl:${versions.log4j}") https://github.com/elastic/elasticsearch/issues/93714
   javaRestTestImplementation("com.wdtinc:mapbox-vector-tile:3.1.0")
   javaRestTestImplementation("com.google.protobuf:protobuf-java:${versions.protobuf}")
-}
 
-testClusters.configureEach {
-  setting 'xpack.license.self_generated.type', 'trial'
-  testDistribution = 'DEFAULT'
-  setting 'xpack.security.enabled', 'false'
+  clusterModules project(':modules:analysis-common')
+  clusterModules project(':modules:legacy-geo')
+  clusterModules project(':modules:lang-painless')
+  clusterModules project(':test:external-modules:test-error-query')
 }
 
 tasks.named("thirdPartyAudit").configure {
